update jumlah barang

+++ database/migrations/2023_04_07_140504_create_produks_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('deskripsi')->nullable();
             $table->integer('stok');
             $table->integer('diskon')->default(0);
-            $table->integer('harga');
+            $table->double('harga');
             $table->timestamps();
         });
     }
